split out TakeoverControls
